+++ R/taxes.R
@@ -2,7 +2,7 @@
 #' @docType data
 #' @title Tax Preference Conjoint Experiment Dataset from Ballard-Rosa et al. (2016)
 #' @description A dataset containing the results of a fully randomized conjoint survey of a representative sample of 2000 American adults who were asked to choose between alternative tax rate policies. Variables \samp{taxrate1}-\samp{taxrate6} refer to tax rates for different income brackets and \samp{taxrev} refers to levels of total tax revenue.
-#' @format A data frame with 32000 observations on the following 13 variables. Each row corresponds to a single profile presented to a respondent.
+#' @format A data frame (with additional \dQuote{cj_df} class) with 32000 observations on the following 13 variables. Each row corresponds to a single profile presented to a respondent.
 #'  \describe{
 #'    \item{\samp{chose_plan}}{A numeric vector denoting whether the immigrant profile was selected (=1) or not (=0).}
 #'    \item{\samp{taxrate1}}{An experimental factor with levels \dQuote{<10k: 0\%}, dQuote{<10k: 5\%}, dQuote{<10k: 15\%}, dQuote{<10k: 25\%}.}
@@ -26,5 +26,5 @@
 #'         taxrate4 + taxrate5 + taxrate6 + taxrev
 #' cj(taxes, f1, id = ~ ID, weights = ~     weight)
 #' }
-#' @seealso \code{\link{cj}} \code{\link{immigration}}
+#' @seealso \code{\link{cj}} \code{\link{immigration}} \code{\link{cj_df}}
 "taxes"